// src/app/entities/plaza-interface.ts
export interface PlazaInterface {
    ID?: number;
    CODIGO_PLAZA?: string;
    CODIGO_EMPLEADO?: string;
    NOMBRE?: string;
    A_PATERNO?: string;
    A_MATERNO?: string;
    ID_DEPENDENCIA?:number;
    CODIGO_DEPENDENCIA?: string;
    DEPENDENCIA?: string;
    ID_UNIDAD?: number;
    UNIDAD?: string;
    CARGO?: string;
    DNI?: string;
    ID_TIPO_CONTRATO?: number;
    ID_ESTADO?: number;
    NOMBRE_ESTADO?: string;
    MODALIDAD?:number;
    XML_REMUNERACION?: string;
  }